backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
PyPDF2==3.0.1
pdf2image==1.17.0
Pillow==10.1.0
werkzeug==3.0.1
SQLAlchemy==2.0.25
python-dotenv==1.0.0
python-docx==1.1.0
